move query knowledge graph function
